AI paths to and melee attacks enemies now
This fixes the bug that let players get a tile of space when running away
But it also introduces an oddity when players run away from AI and the AI is getting its pathing started

Also, we will need to create events that persist for the duration of a turn and the turn afterwards eventually.
It would be good to start counting turns if we do this as well.