Alert system
Enemies will go on alert for a turn once noticing the player.
This gives the player a chance to react to their presence being discovered.
This will be expanded upon further. Being cumblobbed should halt detection progress, but removal of cumblob should instantly reveal the location of the player if they are within sight range.
Still need to add mechanics where enemies have a chance to detect the player with certain actions .
Also need to add perception to the detection chance dice roll. Enemies seem a little bit too unperceptive?
Also need to add chance for enemies to lose track of the player.